Sprite update + click detection
Updated the sprites. Also added collision boxes to a number of the items so we can tell when they're actually being clicked on. The door will open when clicked, clicking on the small numpad next to the door will open up a large version of the numpad, and clicking off of this large numpad will close it. We can change the numpad to a combo lock fairly easily if we want, just would have to change the sprites around again.