Lock on the Door
The door now stays locked until the player enters the correct combo into the numpad. They can do so by clicking on the up/down arrows on screen under each number space. The combo is currently set to 3241, since that is the order the letters on the wall should be in to spell help instead of leph.