Fixed canvas issues, added more sound effects, changed code
The numpad's numbers and arrows now follow the numpad around properly + don't shift when the aspect ratio is changed. The numpad will always appear in the center of the screen no matter where you open it, and you must now click the red x in its top right corner to close it. I also tidied up that code a bit and locked the player's movement while it is open. In addition, sound effects have been added to every object, and I updated the bookshelf sprite to match the rest of the room. The code for the door is now 5479. 5 for the gems under the bed, 4 for the number of white books on the bookshelf, 7 for the blank spaces in the bookshelf, and 9 for the number of blue stripes on the wall. The hints have been updated accordingly.