Instructions + A/D key input
Added the option to use the A/D keys to move across the room in addition to the left/right arrows. Added simple instructions on the bottom of the screen letting the player know that they can move in such a way and click on objects to interact with them. These instructions disappear once the player moves.